+++ src/test/java/org/springframework/samples/petclinic/ui/PossitiveListPropertiesOwner.java
@@ -24,11 +24,79 @@ public class PossitiveListPropertiesOwner {
 
   @BeforeEach
   public void setUp() throws Exception {
-    String pathToGeckoDriver = "C:\\Users\\enriq\\OneDrive\\Escritorio";
-    System.setProperty("webdriver.chrome.driver", pathToGeckoDriver + "\\chromedriver.exe");
+//    String pathToGeckoDriver = "C:\\Users\\enriq\\OneDrive\\Escritorio";
+    System.setProperty("webdriver.chrome.driver",System.getenv("webdriver.chrome.driver"));
     driver = new ChromeDriver();
+   
     baseUrl = "https://www.google.com/";
     driver.manage().timeouts().implicitlyWait(30, TimeUnit.SECONDS);
+  }
+  
+  @Test
+  public void testNegativeCreatePropertyUI() throws Exception {
+    driver.get("http://localhost:8090/");
+    driver.findElement(By.xpath("//a[contains(@href, '/login')]")).click();
+    driver.findElement(By.id("username")).click();
+    driver.findElement(By.id("username")).clear();
+    try {
+        Thread.sleep(500);
+    } catch (InterruptedException e) {
+    }
+    driver.findElement(By.id("username")).sendKeys("owner1");
+    driver.findElement(By.id("password")).clear();
+    try {
+        Thread.sleep(500);
+    } catch (InterruptedException e) {
+    }
+    driver.findElement(By.id("password")).sendKeys("1");
+    driver.findElement(By.xpath("//button[@type='submit']")).click();
+    driver.findElement(By.xpath("//div[@id='main-navbar']/ul/li[2]/a/span[2]")).click();
+    try {
+        Thread.sleep(1000);
+    } catch (InterruptedException e) {
+    }
+  }
+
+  @AfterEach
+  public void tearDown() throws Exception {
+    driver.quit();
+    String verificationErrorString = verificationErrors.toString();
+    if (!"".equals(verificationErrorString)) {
+      fail(verificationErrorString);
+    }
+  }
+
+  private boolean isElementPresent(By by) {
+    try {
+      driver.findElement(by);
+      return true;
+    } catch (NoSuchElementException e) {
+      return false;
+    }
+  }
+
+  private boolean isAlertPresent() {
+    try {
+      driver.switchTo().alert();
+      return true;
+    } catch (NoAlertPresentException e) {
+      return false;
+    }
+  }
+
+  private String closeAlertAndGetItsText() {
+    try {
+      Alert alert = driver.switchTo().alert();
+      String alertText = alert.getText();
+      if (acceptNextAlert) {
+        alert.accept();
+      } else {
+        alert.dismiss();
+      }
+      return alertText;
+    } finally {
+      acceptNextAlert = true;
+    }
   }
 
 }
